refactor(indicator): remove unused ;*(
