+++ packages/indicator/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@alfalab/core-components-indicator",
-    "version": "1.0.2",
+    "version": "0.0.0",
     "description": "",
     "keywords": [],
     "license": "MIT",
