+++ packages/indicator/src/docs/Component.stories.mdx
@@ -31,10 +31,10 @@ import Changelog from '../../CHANGELOG.md';
                 <Indicator
                     value={number('value', 1)}
                     height={number('height', 20)}
-                    color={text('color', 'text-primary-inverted')}
-                    backgroundColor={text('backgroundColor', 'bg-accent')}
+                    color={text('color', 'var(--color-light-text-primary-inverted)')}
+                    backgroundColor={text('backgroundColor', 'var(--color-light-bg-accent)')}
                     border={boolean('border', true) && {
-                        color: text('border.color', 'border-primary-inverted'),
+                        color: text('border.color', 'var(--color-light-border-primary-inverted)'),
                         width: number('border.width', 1),
                     }}
                     view={select('view', [undefined, 'red', 'grey', 'white'], undefined)}
